add delete and update tests

+++ patient/src/main/java/com/nutridata/patient/PatientService.java
@@ -30,6 +30,26 @@ public class PatientService {
     }
 
     public void deletePatient(Long id) {
+        if (id <= 0) {
+            throw new InvalidPatientIdException("Patient ID can not be negative or zero");
+        }
+        if (!patientRepository.existsById(id)) {
+            throw new PatientNotFoundException("Patient with id " + id + " not found");
+        }
+
         this.patientRepository.deleteById(id);
+    }
+
+    public void updatePatient(Long id, PatientDTO patientDTO) {
+        Patient existingPatient = this.getPatient(id);
+
+        if (patientDTO.getName() != null) existingPatient.setName(patientDTO.getName());
+        if (patientDTO.getSurname() != null) existingPatient.setSurname(patientDTO.getSurname());
+        if (patientDTO.getEmail() != null) existingPatient.setEmail(patientDTO.getEmail());
+        if (patientDTO.getAge() != null) existingPatient.setAge(patientDTO.getAge());
+        if (patientDTO.getWeight() != null) existingPatient.setWeight(patientDTO.getWeight());
+        if (patientDTO.getHeight() != null) existingPatient.setHeight(patientDTO.getHeight());
+
+        this.patientRepository.save(existingPatient);
     }
 }
